+++ src/co/multiply/pathling/TransformKeys.java
@@ -25,7 +25,7 @@ public final class TransformKeys {
         if (nav == null) {
             return obj;
         }
-        return ((Nav.Updatable) nav).applyUpdates(obj, tf);
+        return ((Nav.Updatable) nav).applyUpdates(obj, new FunctionReplacer(tf));
     }
 
     // ========================================================================